Update the padding for header class
For considerable and equal sized space on either sides
of the webpage use 80% minimum width and have the
padding on both sides set to 20%.

For some reason, I used shorthand code for margin but
the effect only takes place on the left side of the
webpage. Same thing happens when I cluster my css style
with long forms of the margin.