Update class frame and round styling
I created an empty div with a defined width, height and a
specifit border radius. I tried to fit an image into the
div and made it to inherit the relevant properties of its
container but the corners of the image still showed the
white background.

I have commmented out the styling for the div frame and
gave the class new styles that add the border to the image
and not the image to a set frame.

The webpage wasn't designed to have an image in the frame
I wrote the styles for. I decided to put an image and found
the error.